--- a/common/dal/src/impls/local.rs
+++ b/common/dal/src/impls/local.rs
@@ -45,10 +45,8 @@ impl Local {
     pub fn with_path(root_path: PathBuf) -> Local {
         Local { root: root_path }
     }
-}
 
-impl Local {
-    fn prefix_with_root(&self, path: &str) -> Result<PathBuf> {
+    pub fn prefix_with_root(&self, path: &str) -> Result<PathBuf> {
         let path = normalize_path(&self.root.join(&path));
         if path.starts_with(&self.root) {
             Ok(path)
@@ -120,7 +118,7 @@ impl DataAccessor for Local {
 }
 
 // from cargo::util::path
-fn normalize_path(path: &Path) -> PathBuf {
+pub fn normalize_path(path: &Path) -> PathBuf {
     let mut components = path.components().peekable();
     let mut ret = if let Some(c @ Component::Prefix(..)) = components.peek().cloned() {
         components.next();