--- a/common/planners/src/plan_insert_into.rs
+++ b/common/planners/src/plan_insert_into.rs
@@ -12,7 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-
 use common_datavalues::DataSchemaRef;
 use common_meta_types::MetaId;