--- a/common/streams/src/stream.rs
+++ b/common/streams/src/stream.rs
@@ -12,7 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-
 use common_datablocks::DataBlock;
 use common_exception::Result;