--- a/query/src/configs/config_test.rs
+++ b/query/src/configs/config_test.rs
@@ -76,6 +76,7 @@ storage_type = \"disk\"
 
 [storage.disk]
 data_path = \"\"
+temp_data_path = \"\"
 
 [storage.s3]
 region = \"\"