--- a/query/src/datasources/table/csv/mod.rs
+++ b/query/src/datasources/table/csv/mod.rs
@@ -13,6 +13,5 @@
 //  limitations under the License.
 //
 pub mod csv_table;
-pub mod csv_table_stream;
 #[cfg(test)]
 mod csv_table_test;